ref: move getHasTag to tag.ts
